requirements.txt
fastapi==0.110.2
sentence_transformers==2.7.0
python-dotenv==1.0.1
langchain_community==0.0.34
langchain_openai==0.1.3
langchain==0.1.16
selenium==4.19.0
requests_html==0.10.0
lxml_html_clean==0.1.1
PyMuPDF==1.24.2
pandas==2.2.2
pdfplumber==0.11.0
PyPDF4==1.27.0
chromadb==0.4.24
faiss-cpu==1.8.0
python-pptx==0.6.23
pikepdf==8.15.1
nltk==3.8.1
PyPDF2==3.0.1
youtube-search-python==1.6.6
google-api-python-client==2.129.0